+++ app/build.gradle
@@ -8,12 +8,12 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.appdev.volume_android"
+        applicationId "com.cornellappdev.volume"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
-
+R
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         androidExtensions {
