+++ app/src/main/java/com/cornellappdev/volume/adapters/SavedArticlesAdapter.kt
@@ -1,4 +1,4 @@
-package com.appdev.volume_android.adapters
+package com.cornellappdev.volume.adapters
 
 import android.content.Intent
 import android.graphics.BlurMaskFilter
@@ -12,42 +12,43 @@ import android.widget.TextView
 import androidx.annotation.RequiresApi
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
-import com.appdev.volume_android.MainActivity
-import com.appdev.volume_android.R
-import com.appdev.volume_android.models.Article
+import com.cornellappdev.volume.MainActivity
+import com.cornellappdev.volume.R
+import com.cornellappdev.volume.models.Article
 import com.squareup.picasso.Picasso
-import kotlinx.android.synthetic.main.vertical_article_home_card.view.*
+import kotlinx.android.synthetic.main.saved_article.view.*
+import kotlinx.android.synthetic.main.saved_article.view.article_img_b
+import kotlinx.android.synthetic.main.saved_article.view.article_shout_count_b
+import kotlinx.android.synthetic.main.saved_article.view.article_title_b
+import kotlinx.android.synthetic.main.saved_article.view.desc_holder_b
+import kotlinx.android.synthetic.main.saved_article.view.post_time_b
+import kotlinx.android.synthetic.main.saved_article.view.pub_name_article_layout_b
+import kotlinx.android.synthetic.main.saved_article.view.saved_article_layout
+import kotlinx.android.synthetic.main.saved_article.view.*
 import java.time.Duration
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
+class SavedArticlesAdapter(private val articles: List<Article>) :
+        RecyclerView.Adapter<SavedArticlesAdapter.SavedArticleVH>() {
 
-class BigReadHomeAdapter(private val articles: List<Article>) :
-        RecyclerView.Adapter<BigReadHomeAdapter.BigReadArticleVH>() {
-
-    class BigReadArticleVH(itemView: View) : RecyclerView.ViewHolder(itemView){
-        val pubName: TextView = itemView.vert_card_pub_name
-        val articleTitle : TextView = itemView.article_title_big_read
-        val articleImg : ImageView = itemView.vert_img_view
-        val postTime: TextView = itemView.big_read_card_layout_time
-        val shoutoutCount: TextView = itemView.big_read_card_layout_shoutouts
-        val dot: TextView = itemView.big_read_card_layout_dot
-
-        val layout: ConstraintLayout = itemView.big_read_layout
+    class SavedArticleVH(itemView : View) : RecyclerView.ViewHolder(itemView) {
+        val pubName: TextView = itemView.pub_name_article_layout_b
+        val articleTitle : TextView = itemView.article_title_b
+        val articleImg : ImageView = itemView.article_img_b
+        val postTime: TextView = itemView.post_time_b
+        val shoutoutCount: TextView = itemView.article_shout_count_b
+        val layout: ConstraintLayout = itemView.desc_holder_b
+        val layoutMain: ConstraintLayout = itemView.saved_article_layout
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BigReadArticleVH {
-        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.vertical_article_home_card, parent, false)
-        return BigReadArticleVH(itemView)
-
-    }
-
-    override fun getItemCount(): Int {
-        return articles.size
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int):  SavedArticleVH {
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.saved_article, parent, false)
+        return SavedArticleVH (itemView)
     }
 
     @RequiresApi(Build.VERSION_CODES.O)
-    override fun onBindViewHolder(holder: BigReadArticleVH, position: Int) {
+    override fun onBindViewHolder(holder: SavedArticleVH, position: Int) {
         val currentItem : Article = articles[position]
         holder.articleTitle.text = currentItem.title
         if(currentItem.nsfw == true) {
@@ -56,21 +57,13 @@ class BigReadHomeAdapter(private val articles: List<Article>) :
             val filter = BlurMaskFilter(radius, BlurMaskFilter.Blur.NORMAL)
             holder.articleTitle.paint.maskFilter = filter
         }
-        if(!currentItem.imageURL.isNullOrEmpty()){
+        if(!currentItem.imageURL.isNullOrBlank()) {
+            holder.articleImg.visibility = View.VISIBLE
             Picasso.get().load(currentItem.imageURL).into(holder.articleImg)
-        } else if (!currentItem.publication?.profileImageURL.isNullOrEmpty()) {
-            Picasso.get()
-                    .load(currentItem.publication?.profileImageURL)
-                    .resize(180, 180)
-                    .centerCrop()
-                    .into(holder.articleImg)
         }
-        holder.postTime.text = currentItem.date
-        //getting article posting time and date in phones timeZone
         val format = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
         val datePublished = LocalDateTime.parse(currentItem.date, format)
         val dur = Duration.between(datePublished, LocalDateTime.now())
-        Log.d("TimeStuff", dur.toDays().toString())
         if(dur.toDays() < 1) {
             val hours = dur.toHours()
             holder.postTime.text = "${hours} h ago"
@@ -107,14 +100,16 @@ class BigReadHomeAdapter(private val articles: List<Article>) :
             }
         }
         holder.shoutoutCount.text = currentItem.shoutouts?.toInt().toString() + " shout-outs"
-        holder.pubName.text = currentItem.publication!!.name
+        holder.pubName.text = currentItem.publication?.name
 
-        holder.layout.setOnClickListener{
+        holder.layoutMain.setOnClickListener{
             val intent = Intent(holder.layout.context, MainActivity::class.java)
-            intent.putExtra("article", currentItem)
+            intent.putExtra("article",currentItem)
             holder.layout.context?.startActivity(intent)
         }
     }
 
-
+    override fun getItemCount(): Int {
+        return articles.size
+    }
 }
