+++ app/src/main/java/com/cornellappdev/volume/views/ObservableWebView.kt
@@ -1,11 +1,11 @@
-package com.appdev.volume_android.views
+package com.cornellappdev.volume.views
 
 import android.content.Context
 import android.util.AttributeSet
 import android.util.Log
 import android.view.View
 import android.webkit.WebView
-import com.appdev.volume_android.R
+import com.cornellappdev.volume.R
 import kotlinx.android.synthetic.main.activity_main.view.*
 
 class ObservableWebView @JvmOverloads constructor(
