+++ app/src/main/res/layout/activity_main.xml
@@ -7,18 +7,18 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <com.appdev.volume_android.views.ObservableWebView
+    <com.cornellappdev.volume.views.ObservableWebView
         android:id="@+id/webview"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <com.appdev.volume_android.views.WebviewTop
+    <com.cornellappdev.volume.views.WebviewTop
         android:id="@+id/webview_top"
         app:layout_constraintTop_toTopOf="parent"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
 
-    <com.appdev.volume_android.views.WebviewBottom
+    <com.cornellappdev.volume.views.WebviewBottom
         android:id="@+id/webview_bot"
         app:layout_constraintBottom_toBottomOf="parent"
         android:layout_width="match_parent"
